修正參考&修正康僑會無法檢視的問題(0617)

+++ K12.SizeView.Modules/SizeView.cs
@@ -182,7 +182,7 @@ namespace K12.SizeView.Modules
             newrow.CreateCells(dataGridViewX1);
 
             string name = "" + row[0];
-            long size = int.Parse("" + row["pg_total_relation_size"]);
+            long size = long.Parse("" + row["pg_total_relation_size"]);
 
             TotleSize += size;
